+++ include/mc/OpenMPMCPiCalculator.h
@@ -5,12 +5,13 @@
 #include "mc/MCPiCalculator.h"
 #include "random/RandomBuilder.hpp"
 
-static long long simulateChunk_OpenMP(long long chunkSamples, RNGType rngType, DistType distType);
+static unsigned long long simulateChunk_OpenMP(unsigned long long chunkSamples, RNGType rngType,
+                                               DistType distType);
 
 class OpenMPMCPiCalculator : public IMCPiCalculator {
    public:
-    double estimatePi(long long totalSamples, int threadCount, long long chunkSize, RNGType rngType,
-                      DistType distType) override;
+    long double estimatePi(unsigned long long totalSamples, int threadCount,
+                      unsigned long long chunkSize, RNGType rngType, DistType distType) override;
 };
 
 IMCPiCalculator* createOpenMPCalculator();
